separated floating viz from floating viz management
