+++ packages/libs/eda/src/lib/map/analysis/FloatingVisualization.tsx
@@ -1,0 +1,156 @@
+import { Tooltip } from '@material-ui/core';
+import { Link } from 'react-router-dom';
+import { v4 as uuid } from 'uuid';
+import { AnalysisState, PromiseHookState } from '../../core';
+
+import { AppState, useAppState } from './appState';
+import {
+  ComputationAppOverview,
+  Visualization,
+} from '../../core/types/visualization';
+import { FloatingDiv } from './FloatingDiv';
+import { FullScreenVisualization } from '../../core/components/visualizations/VisualizationsContainer';
+import { GeoConfig } from '../../core/types/geoConfig';
+import { EntityCounts } from '../../core/hooks/entityCounts';
+import { VariableDescriptor } from '../../core/types/variable';
+import { Filter } from '../../core/types/filter';
+import { VisualizationPlugin } from '../../core/components/visualizations/VisualizationPlugin';
+
+interface Props {
+  analysisState: AnalysisState;
+  updateVisualizations: (
+    visualizations:
+      | Visualization[]
+      | ((visualizations: Visualization[]) => Visualization[])
+  ) => void;
+  setActiveVisualizationId: ReturnType<
+    typeof useAppState
+  >['setActiveVisualizationId'];
+  appState: AppState;
+  visualizationPlugins: Partial<Record<string, VisualizationPlugin>>;
+  app: ComputationAppOverview;
+  geoConfigs: GeoConfig[];
+  totalCounts: PromiseHookState<EntityCounts>;
+  filteredCounts: PromiseHookState<EntityCounts>;
+  toggleStarredVariable: (variable: VariableDescriptor) => void;
+  filters: Filter[];
+}
+
+export default function FloatingVisualization({
+  analysisState,
+  appState,
+  updateVisualizations,
+  setActiveVisualizationId,
+  geoConfigs,
+  app,
+  visualizationPlugins,
+  totalCounts,
+  filteredCounts,
+  toggleStarredVariable,
+  filters,
+}: Props) {
+  const activeViz = analysisState.analysis?.descriptor.computations
+    .flatMap((c) => c.visualizations)
+    .find((v) => v.visualizationId === appState.activeVisualizationId);
+
+  const fullScreenActions = (
+    <>
+      <div>
+        <Tooltip title="Delete visualization">
+          <button
+            aria-label={`Delete ${activeViz?.displayName || 'visualization.'}`}
+            type="button"
+            className="link"
+            onClick={() => {
+              if (activeViz == null) return;
+              updateVisualizations((visualizations) =>
+                visualizations.filter(
+                  (v) => v.visualizationId !== activeViz.visualizationId
+                )
+              );
+              setActiveVisualizationId(undefined);
+            }}
+          >
+            <i aria-hidden className="fa fa-trash"></i>
+          </button>
+        </Tooltip>
+      </div>
+      <div>
+        <Tooltip title="Copy visualization">
+          <button
+            aria-label={`Create a copy of ${
+              activeViz?.displayName || 'visualization.'
+            }`}
+            type="button"
+            className="link"
+            onClick={() => {
+              if (activeViz == null) return;
+              const vizCopyId = uuid();
+              updateVisualizations((visualizations) =>
+                visualizations.concat({
+                  ...activeViz,
+                  visualizationId: vizCopyId,
+                  displayName:
+                    'Copy of ' +
+                    (activeViz.displayName || 'unnamed visualization'),
+                })
+              );
+              setActiveVisualizationId(vizCopyId);
+            }}
+          >
+            <i aria-hidden className="fa fa-clone"></i>
+          </button>
+        </Tooltip>
+      </div>
+      <Tooltip title="Minimize visualization">
+        <Link
+          to=""
+          onClick={(e) => {
+            e.preventDefault();
+            setActiveVisualizationId(undefined);
+          }}
+        >
+          <i aria-hidden className="fa fa-window-minimize" />
+        </Link>
+      </Tooltip>
+    </>
+  );
+
+  const computation = analysisState.analysis?.descriptor.computations[0];
+
+  return (
+    <>
+      {activeViz && (
+        <FloatingDiv
+          style={{
+            bottom: 10,
+            left: 500,
+            transformOrigin: 'bottom left',
+            transform: 'scale(0.8)',
+          }}
+        >
+          <FullScreenVisualization
+            analysisState={analysisState}
+            computation={computation!}
+            updateVisualizations={updateVisualizations}
+            visualizationPlugins={visualizationPlugins}
+            visualizationsOverview={app.visualizations}
+            geoConfigs={geoConfigs}
+            computationAppOverview={app}
+            filters={filters}
+            starredVariables={
+              analysisState.analysis?.descriptor.starredVariables ?? []
+            }
+            toggleStarredVariable={toggleStarredVariable}
+            totalCounts={totalCounts}
+            filteredCounts={filteredCounts}
+            isSingleAppMode
+            disableThumbnailCreation
+            id={activeViz.visualizationId}
+            actions={fullScreenActions}
+          />
+        </FloatingDiv>
+      )}
+    </>
+  );
+}
